upload_service/requirements.txt
colorama==0.4.6
hiredis==3.2.1
Kurigram==2.2.10
loguru==0.7.3
pyaes==1.6.1
PySocks==1.7.1
redis==6.4.0
TgCrypto==1.2.5
win32_setctime==1.2.0
python-dotenv==1.0.1
